requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
tensorflow
mysql-connector-python
joblib
jupyter
